Update archive_local_COPY AND EDIT.bat
Allows override of the current directory variable in command shell for those cases where the current directory is set to the windows directory or some generic location other than the default location of the target file. Used most often when calling the batch file from a link in OneNote, because OneNote runs the file from the Windows system directory.